+++ CHANGELOG.md
@@ -11,6 +11,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - `PMG\Queue\Router::queueFor` now typehints against `object` instead of
   `Message`.
 - All `PMG\Queue\MessageLifecycle` methods now typehint against `object`.
+- `PMG\Queue\Envelope` and `DefaultEnvelope` now deal with `object` messages
+  only and do not typehint again `Message`.
 - `MessageLifecycle::failed` no longer has an `$isRetrying` argument, instead
   `MessageLifecycyle::retrying` will be called instead.
 - The `Message` interface no longer has a `getName` method.